src/chap7/Shape.java
/*
 * This code is for the learning of Java
 * It is not, and is not intended to be, production grade code.   * 
 * Use at your own risk.  * 
 */
package chap7;

/**
 *
 * @author steve
 */
public abstract class Shape {
    private String color;
    public Shape(String color){
        this.setColor(color);
    }

    /**
     * @return the color
     */
    public String getColor() {
        return color;
    }

    /**
     * @param color the color to set
     */
    public void setColor(String color) {
        this.color = color;
    }
    public abstract double calculateArea();
}
